Refactored code

+++ pinecone_util.py
@@ -5,26 +5,30 @@ from pinecone import Pinecone
 from pinecone import ServerlessSpec
 
 class PineconeUtil:
-    def __init__(self):
+    def __init__(self, create_index = False):
         self.index_name = 'df-resource'
+        self.create_index = create_index
+
         self.initialize_pinecone()
+
 
     # Initialize Pinecone
     def initialize_pinecone(self):
         load_dotenv()  # Load environment variables from .env file
         self.pinecone = Pinecone(api_key=os.getenv('PINECONE_API_KEY'))
 
-        if self.index_name in self.pinecone.list_indexes().names():
-            print('Deleting existing index')
-            self.pinecone.delete_index(self.index_name)
+        if self.create_index:
+            if self.index_name in self.pinecone.list_indexes().names():
+                print('Deleting existing index')
+                self.pinecone.delete_index(self.index_name)
 
-        print('Creating index...')
-        self.pinecone.create_index(
-            name=self.index_name,
-            dimension=384,  # Correct dimension for the embedding model used
-            metric="cosine",
-            spec=ServerlessSpec(cloud='aws', region='us-east-1')
-        )
+            print('Creating index...')
+            self.pinecone.create_index(
+                name=self.index_name,
+                dimension=384,  # Correct dimension for the embedding model used
+                metric="cosine",
+                spec=ServerlessSpec(cloud='aws', region='us-east-1')
+            )
 
         self.pinecone_index = self.pinecone.Index(self.index_name)
 
@@ -33,6 +37,7 @@ class PineconeUtil:
             time.sleep(1)
 
         print(f"Got index '{self.index_name}'")
+
 
     # Store embeddings in Pinecone
     def store_embeddings(self, embeddings):
@@ -43,10 +48,7 @@ class PineconeUtil:
             batch = vectors[i:i + batch_size]
             self.pinecone_index.upsert(vectors=batch)
 
+
     # Process the query
     def process_query(self, query_embedding):
-        result = self.pinecone_index.query(vector=[query_embedding], top_k=5)
-        
-        #index = Pinecone.from_documents(original_data, query_embedding, index_name = self.index_name)
-        #result = index.similarity_search(query_embedding, k=5)
-        return result
+        return self.pinecone_index.query(vector=[query_embedding], top_k=5)
